Create README.md
This is the simplest of buildings we can generate. A Box-Shaped building, one adjacent building, one window on each side, and an overhang and two fins on the south side. The format of this test case folder will be a .rad file with the corresponding bitmaps for this rad file.